[muzi] adding getRootPath test and code. refactoring create-all-templates :fist:
