--- src/tests/test.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const expect = require('chai').expect;
-
-describe('Dummy Test', () => {
-    it('should pass a dummy test', () => {
-        const result = true;
-        expect(result).to.equal(true);
-    });
-});
